Testing: add function should add a new item to the list array and save it to local storage.
